refactor(ui): move the danger zone from Saves to the General settings tab
The three reset tiers now live in a new general.rs (the General section
previously rendered the placeholder); Saves is back to export/import
only. Verified in-browser: General shows the danger zone, Saves no
longer does.
