+++ crates/synergismforkd_ui/src/sections/settings/mod.rs
@@ -1,4 +1,5 @@
 //! Settings group sections.
 
+pub mod general;
 pub mod saves;
 pub mod themes;
